+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.37] - 2024-12-19
+
+### Added
+- People list: Sort controls to sort by first name or last name, ascending or descending
+- People list: Default sorting is now first name ascending (changed from last name)
+
+### Changed
+- People list: Sorting now uses a dropdown selector and order toggle button
+
 ## [1.0.36] - 2024-12-19
 
 ### Changed